requirements.txt
aiogram>=3.7.0
rembg==2.0.72
Pillow>=11.0.0
python-dotenv==1.1.0
openai>=1.0.0
